Improve logic to determine when to show primary admission menu
We now define a list of 5 site paths that require the primary menu
for admission.wsu.edu. Additional site paths can be added to the
array.

For local development, filter the `admission_home_domain` value
to something like `dev.admission.wsu.edu` and create sites with
similar paths.

If BU Navigation is enabled on the primary site, then it should
also be enabled on any of the other sites. If not, the other sites
will still fall back to the non BU menu from the primary site.